requirements.txt
pytesseract==0.3.0
opencv_python==3.4.2.16
opencv-contrib-python==3.4.2.16
numpy==1.16.3
Pillow==6.2.1